+++ nettle-meta.h
@@ -37,13 +37,13 @@ struct nettle_cipher
 {
   const char *name;
   
-  unsigned context_size;
+  size_t context_size;
   
   /* Zero for stream ciphers */
-  unsigned block_size;
+  size_t block_size;
 
   /* Suggested key size; other sizes are sometimes possible. */
-  unsigned key_size;
+  size_t key_size;
 
   nettle_set_key_func *set_encrypt_key;
   nettle_set_key_func *set_decrypt_key;
